// front-orders/src/modules/dashboard/modules/subscription/enum/subscription.form.enum.ts
export enum SubscriptionFormEnum {
  SUBCRIPTION_DETAIL = 'SUBCRIPTION_DETAIL',
  SUBCRIPTION_RECEIPT = 'SUBCRIPTION_RECEIPT',
}

export enum PaymentMethodEnum {
  ONLINE = 'ONLINE',
  TRANSFER = 'TRANSFER',
}

export enum PaymentStatusEnum {
  PENDING = 'PENDING',
  SUCCESS = 'SUCCESS',
  REJECTED = 'REJECTED',
}

export enum TabInvestmentSubscriptionEnum {
  DETAIL = 0,
  BANKS = 1,
  RECEIPT = 2,
}